moved config to its own package outside of games
