+++ src/main/java/org/braekpo1nt/mctmanager/config/YawPitch.java
@@ -1,4 +1,4 @@
-package org.braekpo1nt.mctmanager.games.game.config;
+package org.braekpo1nt.mctmanager.config;
 
 /**
  * Simple record to store the yaw and pitch of a Location. Useful for serializing the directional portion of a Location in JSON.
