+++ src/main/java/org/braekpo1nt/mctmanager/config/inventory/ItemStackDTO.java
@@ -1,8 +1,8 @@
-package org.braekpo1nt.mctmanager.games.game.config.inventory;
+package org.braekpo1nt.mctmanager.config.inventory;
 
 import com.google.common.base.Preconditions;
-import org.braekpo1nt.mctmanager.games.game.config.enchantments.EnchantmentDTO;
-import org.braekpo1nt.mctmanager.games.game.config.inventory.meta.ItemMetaDTO;
+import org.braekpo1nt.mctmanager.config.enchantments.EnchantmentDTO;
+import org.braekpo1nt.mctmanager.config.inventory.meta.ItemMetaDTO;
 import org.bukkit.Material;
 import org.bukkit.NamespacedKey;
 import org.bukkit.enchantments.Enchantment;
