+++ src/main/java/org/braekpo1nt/mctmanager/games/colossalcombat/config/ColossalCombatStorageUtil.java
@@ -6,8 +6,7 @@ import com.google.gson.JsonSyntaxException;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.serializer.gson.GsonComponentSerializer;
 import org.braekpo1nt.mctmanager.Main;
-import org.braekpo1nt.mctmanager.games.game.config.GameConfigStorageUtil;
-import org.braekpo1nt.mctmanager.games.game.config.inventory.PlayerInventoryDTO;
+import org.braekpo1nt.mctmanager.config.GameConfigStorageUtil;
 import org.bukkit.Bukkit;
 import org.bukkit.Location;
 import org.bukkit.Material;
@@ -20,7 +19,6 @@ import org.jetbrains.annotations.Nullable;
 
 import java.io.File;
 import java.io.InputStream;
-import java.util.Arrays;
 
 public class ColossalCombatStorageUtil extends GameConfigStorageUtil<ColossalCombatConfig> {
