+++ src/main/java/org/braekpo1nt/mctmanager/games/event/config/EventStorageUtil.java
@@ -2,7 +2,7 @@ package org.braekpo1nt.mctmanager.games.event.config;
 
 import com.google.common.base.Preconditions;
 import org.braekpo1nt.mctmanager.Main;
-import org.braekpo1nt.mctmanager.games.game.config.GameConfigStorageUtil;
+import org.braekpo1nt.mctmanager.config.GameConfigStorageUtil;
 import org.jetbrains.annotations.Nullable;
 
 import java.io.File;
