+++ src/main/java/org/braekpo1nt/mctmanager/games/game/capturetheflag/config/LoadoutDTO.java
@@ -4,9 +4,9 @@ import com.google.common.base.Preconditions;
 import com.google.gson.JsonElement;
 import lombok.Getter;
 import org.braekpo1nt.mctmanager.games.game.capturetheflag.Loadout;
-import org.braekpo1nt.mctmanager.games.game.config.ConfigUtil;
-import org.braekpo1nt.mctmanager.games.game.config.inventory.PlayerInventoryDTO;
-import org.braekpo1nt.mctmanager.games.game.config.inventory.meta.ItemMetaDTO;
+import org.braekpo1nt.mctmanager.config.ConfigUtil;
+import org.braekpo1nt.mctmanager.config.inventory.PlayerInventoryDTO;
+import org.braekpo1nt.mctmanager.config.inventory.meta.ItemMetaDTO;
 import org.bukkit.Material;
 import org.bukkit.inventory.ItemStack;
 import org.jetbrains.annotations.NotNull;
