+++ src/main/java/org/braekpo1nt/mctmanager/games/game/clockwork/config/ClockworkConfig.java
@@ -2,10 +2,9 @@ package org.braekpo1nt.mctmanager.games.game.clockwork.config;
 
 
 import com.google.gson.JsonElement;
-import org.braekpo1nt.mctmanager.games.game.config.BoundingBoxDTO;
-import org.braekpo1nt.mctmanager.games.game.config.SoundDTO;
+import org.braekpo1nt.mctmanager.config.BoundingBoxDTO;
+import org.braekpo1nt.mctmanager.config.SoundDTO;
 import org.bukkit.scoreboard.Team;
-import org.bukkit.util.BoundingBox;
 import org.bukkit.util.Vector;
 
 import java.util.List;
