+++ src/main/java/org/braekpo1nt/mctmanager/games/game/footrace/config/FootRaceConfig.java
@@ -1,11 +1,8 @@
 package org.braekpo1nt.mctmanager.games.game.footrace.config;
 
-import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
-import org.braekpo1nt.mctmanager.games.game.config.BoundingBoxDTO;
-import org.braekpo1nt.mctmanager.games.game.config.LocationDTO;
-import org.bukkit.util.BoundingBox;
-import org.bukkit.util.Vector;
+import org.braekpo1nt.mctmanager.config.BoundingBoxDTO;
+import org.braekpo1nt.mctmanager.config.LocationDTO;
 
 record FootRaceConfig(String version, String world, LocationDTO startingLocation, BoundingBoxDTO finishLine, BoundingBoxDTO spectatorArea, BoundingBoxDTO glassBarrier, Scores scores, Durations durations, JsonElement description) {
