+++ src/main/java/org/braekpo1nt/mctmanager/games/game/spleef/config/PowerupDTO.java
@@ -1,8 +1,7 @@
 package org.braekpo1nt.mctmanager.games.game.spleef.config;
 
-import com.google.common.base.Preconditions;
 import lombok.Getter;
-import org.braekpo1nt.mctmanager.games.game.config.SoundDTO;
+import org.braekpo1nt.mctmanager.config.SoundDTO;
 import org.braekpo1nt.mctmanager.games.game.spleef.powerup.Powerup;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
