+++ src/main/java/org/braekpo1nt/mctmanager/games/game/spleef/config/SpleefConfig.java
@@ -3,9 +3,9 @@ package org.braekpo1nt.mctmanager.games.game.spleef.config;
 import com.google.common.base.Preconditions;
 import com.google.gson.JsonElement;
 import lombok.Getter;
-import org.braekpo1nt.mctmanager.games.game.config.BoundingBoxDTO;
-import org.braekpo1nt.mctmanager.games.game.config.NamespacedKeyDTO;
-import org.braekpo1nt.mctmanager.games.game.config.inventory.ItemStackDTO;
+import org.braekpo1nt.mctmanager.config.BoundingBoxDTO;
+import org.braekpo1nt.mctmanager.config.NamespacedKeyDTO;
+import org.braekpo1nt.mctmanager.config.inventory.ItemStackDTO;
 import org.braekpo1nt.mctmanager.games.game.spleef.powerup.Powerup;
 import org.bukkit.Material;
 import org.bukkit.util.Vector;
