+++ src/main/java/org/braekpo1nt/mctmanager/utils/EntityUtils.java
@@ -1,9 +1,8 @@
 package org.braekpo1nt.mctmanager.utils;
 
-import org.braekpo1nt.mctmanager.games.game.config.YawPitch;
+import org.braekpo1nt.mctmanager.config.YawPitch;
 import org.bukkit.Location;
 import org.bukkit.block.BlockFace;
-import org.bukkit.entity.Player;
 import org.bukkit.util.Vector;
 
 public final class EntityUtils {
